Move ReactTrackingContext into separate file, update useTracking tests
