+++ src/__tests__/useTracking.test.js
@@ -1,25 +1,9 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
 import { mount } from 'enzyme';
 import React from 'react';
-import { renderToString } from 'react-dom/server';
-import track from '../withTrackingComponentDecorator';
 import useTracking from '../useTracking';
 
 describe('useTracking', () => {
-  it('throws error if tracking context not present', () => {
-    const ThrowMissingContext = () => {
-      useTracking();
-      return <div>hi</div>;
-    };
-    try {
-      renderToString(<ThrowMissingContext />);
-    } catch (error) {
-      expect(error.message).toContain(
-        'Attempting to call `useTracking` without a ReactTrackingContext present'
-      );
-    }
-  });
-
   it('dispatches tracking events from a useTracking hook tracking object', () => {
     const outerTrackingData = {
       page: 'Page',
@@ -27,8 +11,8 @@ describe('useTracking', () => {
 
     const dispatch = jest.fn();
 
-    const App = track(outerTrackingData, { dispatch })(() => {
-      const tracking = useTracking();
+    const App = () => {
+      const tracking = useTracking(outerTrackingData, { dispatch });
 
       expect(tracking.getTrackingData()).toEqual({
         page: 'Page',
@@ -44,7 +28,7 @@ describe('useTracking', () => {
           }
         />
       );
-    });
+    };
 
     const wrapper = mount(<App />);
     wrapper.simulate('click');
